2021/2021-01-19_ado-wiki/requirements.txt
azure-identity==1.5.0
black==20.8b1
hydra-core==1.0.5
jmespath==0.10.0
requests==2.25.1
